requirements/base.txt
beautifulsoup4
celery
dj-database-url
Django
django-test-without-migrations
lxml_to_dict
mysqlclient
PyPDF2
python-decouple
pytz
requests
zeep
